[Enhancment] Now possible to get SRT version of subtitle

+++ Content Server/pages/api/subtitles/get.js
@@ -27,6 +27,7 @@ export default (req, res) => {
         let type = req.query.type;
         let subtitleID = req.query.id;
         let offset = req.query.start ? req.query.start : 0;
+        let useSRT = req.query.useSRT ? true : false;
         let minutes = Math.floor(offset / 60);
         let hours = Math.floor(minutes / 60);
         minutes -= hours * 60;
@@ -58,30 +59,43 @@ export default (req, res) => {
         console.log(output);
 
         // Transcode the subtitle to vtt, with offset according the variables hours, minutes and seconds
-        ffmpeg(subPath)
-        .outputOptions([
-            `-ss ${hours}:${minutes}:${seconds}`
-        ])
-        .output(output)
-        .on('end', function() {
-            // Read the new subtitle file and send it to the user
-            fs.readFile(output, (err, data) => {
+        if (useSRT) {
+            console.log("Sending SRT directly");
+            fs.readFile(subPath, (err, data) => {
                 if (err) {
                     console.log(err);
                     return;
                 }
                 res.status(200).send(data);
-                // Remove the new subtitle file
-                fs.unlink(output, () => {});
                 resolve();
             });
-        })
-        .on('error', (err, stdout, stderr) => {
-            console.log(err);
-            console.log(stdout);
-            console.log(stderr);
-        })
-        .run()
+        } else {
+            ffmpeg(subPath)
+            .outputOptions([
+                `-ss ${hours}:${minutes}:${seconds}`
+            ])
+            .output(output)
+            .on('end', function() {
+                // Read the new subtitle file and send it to the user
+                fs.readFile(output, (err, data) => {
+                    if (err) {
+                        console.log(err);
+                        return;
+                    }
+                    res.status(200).send(data);
+                    // Remove the new subtitle file
+                    fs.unlink(output, () => {});
+                    resolve();
+                });
+            })
+            .on('error', (err, stdout, stderr) => {
+                console.log(err);
+                console.log(stdout);
+                console.log(stderr);
+            })
+            .run()
+        }
+
     });
     
 }
